test/websocket_client/requirements.txt
websocket-client==1.7.0
protobuf==4.25.3
